feature: maps geolocation migrated to backend
